Implemented per-texture anisotropic filtering plus engine-side controls to configure behaviour (cvar: rend-tex-filter-anisotropic and added to the textures tab in the control panel. Replaced the ccmd -anifilter with -noanifilter which can be used to disable anisotropic filtering completely within drOpenGL.
Further minor tweaks to WIN32-specific stuff.

+++ doomsday/plugins/opengl/portable/include/dropengl.h
@@ -101,7 +101,7 @@ extern boolean  noArrays;
 extern int      verbose;
 extern int      useAnisotropic;
 extern int      useVSync;
-extern float    maxAniso;
+extern int      maxAniso;
 extern int      maxTexUnits;
 extern boolean  wireframeMode;
